Clarity in shard note
The notes about sharding is misleading and doesn't clarify the usage of the property. We'll assume if they don't understand this property they are not sharding and set this value to 1 explicitly.

Sharding can be a complex system and is done differently by every developer and as such there is no specific way to recommend what to do with `TotalShards`. This note just clarifies that you set the total shards to your total shards of the bot or set to 1 if they don't understand sharding. They will understand this value when they start sharding their own bot and become more aware of it's usage.